fix: Call Resolve error from calling viewmodel before the fragment is attached
