+++ app/src/main/java/com/example/letschat/ui/viewmodels/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.letschat.Login
+package com.example.letschat.ui.viewmodels
 
 
 import com.example.letschat.repository.DeviceRepository
