+++ app/src/main/res/layout/fragment_login.xml
@@ -4,7 +4,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Login.LoginFragment">
+        tools:context=".ui.fragments.LoginFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
