+++ app/src/main/res/layout/fragment_splash.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SplashFragment">
+    tools:context=".ui.fragments.SplashFragment">
 
    <ImageView
        android:layout_width="wrap_content"
